💄 Design: 챌린지 진행화면 본문 UI 완성
