+++ src/pages/challengePage/components/style/OngoingChallenge.module.scss
@@ -2,14 +2,15 @@
 @use '../../../../styles/variables' as *;
 
 .todayChallenge {
+  @include m.flex-column-center(flex-start, center, 6px);
   width: 100%;
   height: 672px;
   background-color: $main-color2;
   border-radius: 12px 12px 0 0;
+  padding-top: 20px;
 
   .challengeTop {
     @include m.flex-column-center(center, center, 8px);
-    background-color: green;
 
     p {
       @include m.font-size-weight(12px, 400);
@@ -31,12 +32,22 @@
     }
   }
 
+  .todayPhotoGrid {
+    width: 327px;
+    height: 292px;
+
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    grid-row-gap: 20px;
+    grid-column-gap: 12px;
+  }
+
   .challengeButtom {
     @include m.flex-column-center(center, center, 23px);
-    background-color: rgb(156, 121, 57);
 
     .twoButton {
       @include m.flex-row-center(center, center, 12px);
+      margin-top: 18px;
     }
 
     .moneyInfo {
